Fixad bugg i component system

+++ headers/actor-component-system/components/transform.h
@@ -2,6 +2,6 @@
 
 struct Transform {
     float scale;
-    float rotation;
+    double rotation;
     math::vector::Vec2<float> position; 
 };
